refactor: updated to compile all dependencies
BREAKING CHANGES: moved some files around to format for lib folder
