+++ .gitignore
@@ -2,6 +2,7 @@ packages/*/coverage
 packages/*/dist
 node_modules
 *.log
+lib
 .vscode
 .DS_Store
 .idea
