+++ packages/authorizations-core/package.json
@@ -2,7 +2,8 @@
   "name": "@availity/authorizations-core",
   "version": "1.0.5",
   "description": "Base configurations for Availity Authorizations",
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "module": "src/index.js",
   "author": "Kasey Powers <kasey.powers@availity.com>",
   "license": "MIT",
   "publishConfig": {
