+++ packages/dl-axios/package.json
@@ -2,7 +2,8 @@
   "name": "@availity/dl-axios",
   "version": "3.0.17",
   "description": "Utility to download files from services",
-  "main": "index.js",
+  "main": "lib/index.js",
+  "module": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
