+++ packages/dl-core/package.json
@@ -1,26 +1,27 @@
 {
-	"name": "@availity/dl-core",
-	"version": "3.0.17",
-	"description": "Utility to download files from services",
-	"main": "src/download.js",
-	"keywords": [
-		"download",
-		"file",
-		"availity"
-	],
-	"author": "Robert McGuinness",
-	"license": "MIT",
-	"peerDependencies": {
-		"@availity/api-core": "^4.0.1"
-	},
-	"devDependencies": {
-		"@availity/api-core": "^6.0.3"
-	},
-	"dependencies": {
-		"js-file-download": "^0.4.7"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"gitHead": "21b480acb7fc70e240d88451cee8adecb5ddde56"
+  "name": "@availity/dl-core",
+  "version": "3.0.17",
+  "description": "Utility to download files from services",
+  "main": "lib/download.js",
+  "module": "src/download.js",
+  "keywords": [
+    "download",
+    "file",
+    "availity"
+  ],
+  "author": "Robert McGuinness",
+  "license": "MIT",
+  "peerDependencies": {
+    "@availity/api-core": "^4.0.1"
+  },
+  "devDependencies": {
+    "@availity/api-core": "^6.0.3"
+  },
+  "dependencies": {
+    "js-file-download": "^0.4.7"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "gitHead": "21b480acb7fc70e240d88451cee8adecb5ddde56"
 }
