+++ packages/env-var/package.json
@@ -2,7 +2,8 @@
   "name": "@availity/env-var",
   "version": "1.9.2",
   "description": "Availity-specific way to determine variables based on the current environment the code is running in",
-  "main": "index.js",
+  "main": "src/index.js",
+  "module": "lib/index.js",
   "types": "index.d.ts",
   "repository": {
     "type": "git",
