+++ packages/message-core/package.json
@@ -3,6 +3,7 @@
   "version": "4.0.3",
   "description": "availity wrapper for postMessage logic",
   "main": "src/index.js",
+  "module": "lib/index.js",
   "types": "types/index.d.ts",
   "author": "Kasey Powers <kasey.powers@availity.com>",
   "license": "MIT",
