+++ packages/native-form/src/flattenObject.js
