+++ packages/resolve-url/package.json
@@ -2,8 +2,9 @@
   "name": "@availity/resolve-url",
   "version": "1.1.1",
   "description": "Resolve absolute url from relative urls",
-  "main": "resolve-url.js",
-  "types": "resolve-url.d.ts",
+  "main": "src/index.js",
+  "module": "lib/index.js",
+  "types": "types/index.d.ts",
   "keywords": [
     "url",
     "relative"
