+++ packages/resolve-url/src/index.js
@@ -1,0 +1,7 @@
+import resolveUrl from './resolve-url';
+import isAbsoluteUrl from './is-absolute-url';
+import relativeToAbsolute from './relative-to-absolute';
+
+export default resolveUrl;
+
+export { isAbsoluteUrl, relativeToAbsolute };
