+++ packages/resolve-url/src/is-absolute-url.js
