+++ packages/resolve-url/test/resolve-url.test.js
@@ -1,4 +1,4 @@
-import resolveUrl from '.';
+import resolveUrl from '../src';
 
 describe('resolve-url', () => {
   beforeEach(() => {
