+++ packages/resolve-url/types/index.d.ts
@@ -1,0 +1,7 @@
+import resolveUrl from './resolve-url.d.ts';
+import isAbsoluteUrl from './is-absolute-url.d.ts';
+import relativeToAbsolute from './relative-to-absolute.d.ts';
+
+export default resolveUrl;
+
+export { isAbsoluteUrl, relativeToAbsolute };
