+++ packages/resolve-url/types/is-absolute-url.d.ts
