+++ packages/yup/package.json
@@ -2,7 +2,8 @@
   "name": "@availity/yup",
   "version": "1.3.1",
   "description": "Additional Methods for Yup validation library",
-  "main": "index.js",
+  "main": "src/index.js",
+  "module": "lib/index.js",
   "typings": "typings/index.d.ts",
   "scripts": {
     "test": "test"
