+++ packages/yup/tests/npi.test.js
@@ -1,5 +1,5 @@
 import * as yup from 'yup';
-import '..';
+import '../src';
 
 describe('npi', () => {
   test('should be valid for empty input', async () => {
@@ -41,7 +41,7 @@ describe('npi', () => {
     const schema = yup.number().npi();
     const valid = await schema.isValid('i2eh56789o');
     expect(valid).toBe(false);
-    });
+  });
 
   test('should be invalid if NPI is not 10 digits in length', async () => {
     const schema = yup.number().npi();
@@ -73,8 +73,6 @@ describe('npi', () => {
   test('should render custom error message', async () => {
     const schema = yup.number().npi('Test Error Message');
 
-    await expect(schema.validate('123')).rejects.toThrow(
-      'Test Error Message'
-    );
+    await expect(schema.validate('123')).rejects.toThrow('Test Error Message');
   });
 });
